requirements.txt
flask==3.1.1
requests==2.32.4
python-pptx==1.0.2
PyPDF2==3.0.1
Pillow==11.3.0 